Paper Frame Info Block
