--- examples/drawings/paper-frame.html
@@ -1,143 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <title>OpenPlan Paper Frame</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="description" content="OpenGeometry Demo">
-  <meta name="author" content="Vishwajeet Mane">
-  <style>
-    body {
-      margin: 0;
-      overflow: hidden;
-    }
-    #app {
-      width: 100%;
-      height: 100vh;
-    }
-  </style>
-</head>
-<body>
-  <div id="app" style="width: 100%; height: 100vh;">
-  </div>
-  <script type="module">
-    import { OpenPlans } from './../../src/index.ts';
-    import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
-
-    let openPlans;
-    let infoBlock;
-    let paperFrame;
-
-    async function init() {
-      const container = document.getElementById('app');
-      openPlans = new OpenPlans(container);
-      await openPlans.setupOpenGeometry();
-
-      openPlans.pencil.mode = "select";
-
-      const door = openPlans.door();
-      door.position.set(-2, 0.01, 0);
-
-      const window = openPlans.window();
-      window.position.set(2, 0.01, 0);
-
-      const wall = openPlans.wall();
-      wall.position.set(0, 0.01, 0);
-
-      paperFrame = openPlans.paperFrame();
-      paperFrame.format = 'A4';
-      paperFrame.margin = 10;
-      paperFrame.paperName = 'VerticalPaper';
-
-      infoBlock = openPlans.infoBlock();
-      
-      const logoInfoBlock = openPlans.rowInfoBlock(
-      [
-        {
-          type: 'logo',
-          width: 8,
-          height: 1,
-          image: 'https://raw.githubusercontent.com/OpenGeometry-io/.github/refs/heads/main/profile/opengeometryTextLogo.png',
-          name: 'OpenGeometry Logo',
-        }
-      ]);
-      
-      const headerBlock = openPlans.rowInfoBlock([
-        {
-          type: 'text',
-          width: 8,
-          height: 1,
-          text: 'OpenPlans Paper Frame Example',
-          fontSize: 2,
-          fontColor: 0x000000,
-        }
-      ]);
-      const titleBlock = openPlans.rowInfoBlock([
-        {
-          type: 'text',
-          width: 8,
-          height: 1,
-          text: 'Build Anything',
-          fontSize: 3,
-          fontColor: 0x44460FF,
-        }
-      ]);
-
-      infoBlock.blockLayout = {
-        layout: `
-          logoInfoBlock
-          headerBlock
-          titleBlock
-        `,
-        blocks: {
-          logoInfoBlock,
-          headerBlock,
-          titleBlock
-        }
-      }
-
-      paperFrame.addBlock(infoBlock);
-    }
-
-    function setUI() {
-      const gui = new GUI();
-      const frameFolder = gui.addFolder('Paper Frame');
-      const frameControls = {
-        'format': 'A4',
-        'orientation': 'portrait',
-        'margin': 10,
-      };
-
-      const paperFrames = openPlans.getEntitiesByType('paperFrame');
-      paperFrames.forEach((paperFrame) => {
-        const subPaperFrame = frameFolder.addFolder(paperFrame.options.name);
-        
-        subPaperFrame.add(frameControls, 'format', ['A4', 'A3', 'A2']).name('Format').onChange((value) => {
-          paperFrame.format = value;
-        });
-        subPaperFrame.add(frameControls, 'orientation', ['portrait', 'landscape']).name('Orientation').onChange((value) => {
-          paperFrame.orientation = value;
-        });
-        subPaperFrame.add(frameControls, 'margin', 0, 50).name('Margin').onChange((value) => {
-          paperFrame.margin = value;
-        });
-      });
-
-      const infoBlockFolder = gui.addFolder('Info Block');
-      const infoBlockControls = {
-        'placement': 'bottomRight',
-      }
-      infoBlockFolder.add(infoBlockControls, 'placement', ['topLeft', 'topRight', 'bottomLeft', 'bottomRight']).name('Placement').onChange((value) => {
-        infoBlock.placement = value;
-        paperFrame.removeBlock(infoBlock.options.id);
-        paperFrame.addBlock(infoBlock);
-      });
-    }
-
-    (async function () {
-      await init();
-      setUI();
-    })();
-  </script>
-</body>
-</html>
