+++ src/drawing/paper-frame.ts
@@ -1,8 +1,5 @@
 import { Polygon, Vector3D } from "../../kernel/dist";
 import * as THREE from 'three';
-import { InfoBlock, LayoutOptions } from "./info-block";
-import { LogoInfoBlock } from "./logo-block";
-// import { RowInfoBlock } from "./row-info-block2";
 
 export type PaperFormat = 'A4' | 'A3' | 'A2' | 'Custom';
 export type PaperOrientation = 'portrait' | 'landscape';
@@ -32,8 +29,6 @@ export class PaperFrame extends Polygon {
 
   private options: PaperFrameOptions;
   private subNodes: Map<string, THREE.Object3D> = new Map();
-
-  private blocks: InfoBlock[] = [];
   
   private readonly Y_OFFSET = 0.0010; // Offset to avoid z-fighting
