+++ src/index.ts
@@ -15,9 +15,8 @@ import * as THREE from 'three';
 import { BaseWall2 } from './elements/base-wall-2';
 import { Event } from './utils/event';
 import { PaperFrame } from './drawing';
-import { InfoBlock } from './drawing/info-block';
-import { LogoInfoBlock } from './drawing/logo-block';
-import { RowInfoBlock, RowInfoBlockOptions } from './drawing/row-info-block2';
+import { LogoInfoBlock, LogoInfoBlockOptions } from './drawing/logo-info-block';
+import { RowInfoBlock, RowInfoBlockOptions } from './drawing/row-info-block';
 
 export class OpenPlans {
   private container: HTMLElement
@@ -394,8 +393,9 @@ export class OpenPlans {
   //   return infoBlock
   // }
 
-  logoInfoBlock() {
-    const logoBlock = new LogoInfoBlock()
+  logoInfoBlock(options:LogoInfoBlockOptions) {
+    const logoBlock = new LogoInfoBlock(options);
+    this.openThree.scene.add(logoBlock);
     return logoBlock
   }
